startupイベントを使ってウェブサーバー起動時にテーブルを作成します。.env.exampleファイルを追加しました
