+++ .env.example
@@ -1,3 +1,4 @@
+
 # shellcheck shell=ksh
 # 上記コメントでShellCheckはスクリプトを正しく解釈し、適切なチェックを実行する
 # Example database URL
