+++ __init__.py
@@ -1,1 +1,1 @@
-# ディレクトリをPythonパッケージとして認識させるために使用するファイル
+# このファイルを含むディレクトリをパッケージとして認識されるようにするためのファイル
