+++ main.py
@@ -1,12 +1,10 @@
-from fastapi import FastAPI,Depends
-from sqlalchemy.orm import Session
-# from .settings import engine, Base, SessionLocal
-from .models import Task,Base
-from .database import engine
-# FastAPIアプリケーションの作成
+from fastapi import FastAPI
+from .settings import engine, Base 
 app = FastAPI()
 
-
+@app.get("/")
+def read_root():
+    return {"message": "Hello, World!"}
 
 # FastAPIのStartupイベントでテーブルを作成する
 @app.on_event("startup")
@@ -30,21 +28,6 @@ def get_db():
 def read_root():
     return {"message": "Hello, World!"}
 
-# # タスク作成エンドポイントの定義
-# @app.post("/tasks/")
-# async def create_task(title: str, description: str, db: Session = Depends(get_db)):
-#     # Taskオブジェクトを作成してセッションに追加
-#     db_task = Task(title=title, description=description)
-#     db.add(db_task)
-#     db.commit()
-#     db.refresh(db_task)
-#     return db_task
-
-# # タスク一覧取得エンドポイントの定義
-# @app.get("/tasks/")
-# async def read_tasks(skip: int = 0, limit: int = 10, db: Session = Depends(get_db)):
-#     tasks = db.query(Task).offset(skip).limit(limit).all()
-#     return tasks
 
 # Pythonスクリプトが直接実行された場合
 if __name__ == "__main__":
