+++ requirements.txt
@@ -3,3 +3,4 @@ uvicorn==0.25.0
 SQLAlchemy==2.0.30
 psycopg2-binary==2.9.9
 python-dotenv== 1.0.1
+
